requirements.txt
alembic==1.16.4
blinker==1.9.0
click==8.2.1
defusedxml==0.7.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
fpdf2==2.8.3
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
PyMySQL==1.1.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.1
Werkzeug==3.1.3
